fix: update calc to big decimal and big integer

+++ backend/src/main/java/com/wesleybritovlk/souls_calculator_api/app/games/demonsouls/dto/DemonSoulsResponse.java
@@ -1,5 +1,7 @@
 package com.wesleybritovlk.souls_calculator_api.app.games.demonsouls.dto;
 
+import java.math.BigInteger;
+
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
@@ -10,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class DemonSoulsResponse {
     @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
     public record Souls(
-            Long soulsNext) {
+            BigInteger soulsNext) {
     }
 }
